Bases and RelationShipMeta and NodeMeta on PropertyManagerMeta
This is a first measurement to reduce runtime imports.
Obviously it also aims to reduce boilerplate code.
The ultimate goal is to make the code less complicated and make
enhancements easier.
