+++ neomodel/properties.py
@@ -1,7 +1,6 @@
 import functools
 import json
 import sys
-import types
 import re
 import uuid
 import warnings
@@ -11,101 +10,11 @@ from datetime import date, datetime
 import pytz
 
 from neomodel import config
-from neomodel.exceptions import InflateError, DeflateError, RequiredProperty
+from neomodel.exceptions import InflateError, DeflateError
 
 
 if sys.version_info >= (3, 0):
     unicode = str
-
-
-def display_for(key):
-    def display_choice(self):
-        return getattr(self.__class__, key).choices[getattr(self, key)]
-    return display_choice
-
-
-class PropertyManager(object):
-    """
-    Common methods for handling properties on node and relationship objects.
-    """
-
-    def __init__(self, **kwargs):
-        properties = getattr(self, "__all_properties__", None)
-        if properties is None:
-            properties = \
-                self.defined_properties(rels=False, aliases=False).items()
-        for name, property in properties:
-            if kwargs.get(name) is None:
-                if getattr(property, 'has_default', False):
-                    setattr(self, name, property.default_value())
-                else:
-                    setattr(self, name, None)
-            else:
-                setattr(self, name, kwargs[name])
-
-            if getattr(property, 'choices', None):
-                setattr(self, 'get_{}_display'.format(name),
-                        types.MethodType(display_for(name), self))
-
-            if name in kwargs:
-                del kwargs[name]
-
-        aliases = getattr(self, "__all_aliases__", None)
-        if aliases is None:
-            aliases = self.defined_properties(
-                aliases=True, rels=False, properties=False).items()
-        for name, property in aliases:
-            if name in kwargs:
-                setattr(self, name, kwargs[name])
-                del kwargs[name]
-
-        # undefined properties (for magic @prop.setters etc)
-        for name, property in kwargs.items():
-            setattr(self, name, property)
-
-    @property
-    def __properties__(self):
-        from .relationship_manager import RelationshipManager
-
-        return dict((name, value) for name, value in vars(self).items()
-                    if not name.startswith('_')
-                    and not callable(value)
-                    and not isinstance(value,
-                                       (RelationshipManager, AliasProperty,))
-                    )
-
-    @classmethod
-    def deflate(cls, properties, obj=None, skip_empty=False):
-        # deflate dict ready to be stored
-        deflated = {}
-        for name, property \
-                in cls.defined_properties(aliases=False, rels=False).items():
-            db_property = property.db_property or name
-            if properties.get(name) is not None:
-                deflated[db_property] = property.deflate(properties[name], obj)
-            elif property.has_default:
-                deflated[db_property] = property.deflate(
-                    property.default_value(), obj
-                )
-            elif property.required or property.unique_index:
-                raise RequiredProperty(name, cls)
-            elif not skip_empty:
-                deflated[db_property] = None
-        return deflated
-
-    @classmethod
-    def defined_properties(cls, aliases=True, properties=True, rels=True):
-        from .relationship_manager import RelationshipDefinition
-        props = {}
-        for baseclass in reversed(cls.__mro__):
-            props.update(dict(
-                (name, property) for name, property in vars(baseclass).items()
-                if (aliases and isinstance(property, AliasProperty))
-                or (properties and isinstance(property, Property)
-                    and not isinstance(property, AliasProperty))
-                or (rels and isinstance(property, RelationshipDefinition))
-            ))
-        return props
 
 
 def validator(fn):
