+++ scripts/neomodel_install_labels
@@ -5,7 +5,7 @@ import sys
 from importlib import import_module
 from argparse import ArgumentParser
 
-from neomodel import db, install_all_labels
+from neomodel.db import client, install_all_labels
 
 
 def load_python_module_or_file(name):
@@ -59,7 +59,7 @@ def main():
 
     # Connect after to override any code in the module that may set the connection
     print('Connecting to {}\n'.format(bolt_url))
-    db.set_connection(bolt_url)
+    client.set_connection(bolt_url)
 
     install_all_labels()
