+++ test/conftest.py
@@ -2,7 +2,8 @@ from __future__ import print_function
 import warnings
 import os
 
-from neomodel import config, db, clear_neo4j_database, change_neo4j_password
+from neomodel import config, clear_neo4j_database, change_neo4j_password
+from neomodel.db import client
 from neo4j.v1 import CypherError
 
 warnings.simplefilter('default')
@@ -11,12 +12,12 @@ config.DATABASE_URL = os.environ.get('NEO4J_BOLT_URL', 'bolt://neo4j:neo4j@local
 config.AUTO_INSTALL_LABELS = True
 
 try:
-    clear_neo4j_database(db)
+    clear_neo4j_database(client)
 except CypherError as ce:
     # handle instance without password being changed
     if 'The credentials you provided were valid, but must be changed before you can use this instance' in str(ce):
-        change_neo4j_password(db, 'test')
-        db.set_connection('bolt://neo4j:test@localhost:7687')
+        change_neo4j_password(client, 'test')
+        client.set_connection('bolt://neo4j:test@localhost:7687')
 
         print("New database with no password set, setting password to 'test'")
         print("Please 'export NEO4J_BOLT_URL=bolt://neo4j:test@localhost:7687' for subsequent test runs")
